Registers
AF, consisting of A, the accumulator register; and F, an internal register that holds the flags.

BC, a general purpose register

DE, a general purpose register

HL, a general purpose register and it has a large number of instructions that simplifies iterative code, like loops; 16-bit arithmetic; and data loading, bit manipulation, and more.

SP, the stack pointer. Used for call stacks and makes it possible for the CPU to natively support function calls and return values. Has a number of specialized instructions that makes it easier to do this.

PC, the Program Counter. Used to track the location of the next executing instruction. Refers to a memory address.